pyproject.toml
[project]
name = "agents-lab"
version = "0.1.0"
description = "ChatBI exploration and experiments for intelligent data querying"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "mysql-connector-python>=8.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
